Move: ChannelTalkProvider
