+++ src/App.js
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import { CookiesProvider } from "react-cookie";
 import { RecoilRoot } from "recoil";
 import ScrollRestoration from "tools/ScrollRestoration";
-import ChannelTalk from "components/Skeleton/ChannelTalk";
+import ChannelTalkProvider from "components/Skeleton/ChannelTalkProvider";
 import I18nextProvider from "lang/I18nextProvider";
 import AlertProvider from "components/Skeleton/AlertProvider";
 import Skeleton from "components/Skeleton/Skeleton";
@@ -27,7 +27,7 @@ const App = () => {
       <RecoilRoot>
         <Router>
           <ScrollRestoration />
-          <ChannelTalk />
+          <ChannelTalkProvider />
           <I18nextProvider />
           <AlertProvider />
           <Skeleton>
